use new api handler for compatible series
