+++ frontend/src/metabase/dashboard/components/AddSeriesModal/AddSeriesModal.jsx
@@ -4,7 +4,6 @@ import PropTypes from "prop-types";
 import { t } from "ttag";
 import { getIn } from "icepick";
 import { connect } from "react-redux";
-import { createSelector } from "reselect";
 import _ from "underscore";
 
 import Visualization from "metabase/visualizations/components/Visualization";
@@ -12,22 +11,9 @@ import Visualization from "metabase/visualizations/components/Visualization";
 import * as MetabaseAnalytics from "metabase/lib/analytics";
 import { color } from "metabase/lib/colors";
 
-import Questions from "metabase/entities/questions";
-import { getMetadataWithHiddenTables } from "metabase/selectors/metadata";
 import { loadMetadataForQueries } from "metabase/redux/metadata";
 
-import { getVisualizationRaw } from "metabase/visualizations";
-import Question from "metabase-lib/Question";
-
 import { QuestionList } from "./QuestionList";
-
-const getQuestions = createSelector(
-  [getMetadataWithHiddenTables, (_state, props) => props.questions],
-  (metadata, questions) =>
-    questions && questions.map(card => new Question(card, metadata)),
-);
-
-// TODO: rework this so we don't have to load all cards up front
 
 class AddSeriesModal extends Component {
   constructor(props, context) {
@@ -36,14 +22,12 @@ class AddSeriesModal extends Component {
     this.state = {
       error: null,
       series: props.dashcard.series || [],
-      badQuestions: {},
       isLoadingMetadata: false,
     };
   }
 
   static propTypes = {
     dashcard: PropTypes.object.isRequired,
-    questions: PropTypes.array,
     dashcardData: PropTypes.object.isRequired,
     fetchCardData: PropTypes.func.isRequired,
     fetchDatabaseMetadata: PropTypes.func.isRequired,
@@ -52,71 +36,36 @@ class AddSeriesModal extends Component {
   };
   static defaultProps = {};
 
-  handleQuestionSelectedChange = async (question, selected) => {
+  handleQuestionSelectedChange = async (card, selected) => {
     const { dashcard, dashcardData } = this.props;
-    const { visualization } = getVisualizationRaw([{ card: dashcard.card }]);
-    const card = question.card();
-    try {
-      if (selected) {
-        if (getIn(dashcardData, [dashcard.id, card.id]) === undefined) {
-          this.setState({ state: "loading" });
-          await this.props.fetchCardData(card, dashcard, {
-            reload: false,
-            clear: true,
-          });
-        }
-        const sourceDataset = getIn(this.props.dashcardData, [
-          dashcard.id,
-          dashcard.card.id,
-        ]);
-        const seriesDataset = getIn(this.props.dashcardData, [
-          dashcard.id,
-          card.id,
-        ]);
-        if (
-          visualization.seriesAreCompatible(
-            { card: dashcard.card, data: sourceDataset.data },
-            { card: card, data: seriesDataset.data },
-          )
-        ) {
-          this.setState({
-            state: null,
-            series: this.state.series.concat(card),
-          });
 
-          MetabaseAnalytics.trackStructEvent(
-            "Dashboard",
-            "Add Series",
-            card.display + ", success",
-          );
-        } else {
-          this.setState({
-            state: "incompatible",
-            badQuestions: { ...this.state.badQuestions, [card.id]: true },
-          });
-          setTimeout(() => this.setState({ state: null }), 2000);
-
-          MetabaseAnalytics.trackStructEvent(
-            "Dashboard",
-            "Add Series",
-            question.type() + ", " + card.display + ", fail",
-          );
-        }
-      } else {
-        this.setState({
-          series: this.state.series.filter(c => c.id !== card.id),
-        });
-
-        MetabaseAnalytics.trackStructEvent("Dashboard", "Remove Series");
-      }
-    } catch (e) {
-      console.error("AddSeriesModal handleQuestionChange", e);
+    if (!selected) {
       this.setState({
-        state: "incompatible",
-        badQuestions: { ...this.state.badQuestions, [card.id]: true },
+        series: this.state.series.filter(c => c.id !== card.id),
       });
-      setTimeout(() => this.setState({ state: null }), 2000);
+
+      MetabaseAnalytics.trackStructEvent("Dashboard", "Remove Series");
+      return;
     }
+
+    if (getIn(dashcardData, [dashcard.id, card.id]) === undefined) {
+      this.setState({ state: "loading" });
+      await this.props.fetchCardData(card, dashcard, {
+        reload: false,
+        clear: true,
+      });
+    }
+
+    this.setState({
+      state: null,
+      series: this.state.series.concat(card),
+    });
+
+    MetabaseAnalytics.trackStructEvent(
+      "Dashboard",
+      "Add Series",
+      card.display + ", success",
+    );
   };
 
   handleRemoveSeries(card) {
@@ -144,17 +93,9 @@ class AddSeriesModal extends Component {
   };
 
   render() {
-    const { dashcard, dashcardData, questions } = this.props;
-    const { badQuestions } = this.state;
+    const { dashcard, dashcardData } = this.props;
 
-    const { visualization } = getVisualizationRaw([{ card: dashcard.card }]);
-
-    const error = this.state.error;
-
-    const enabledQuestions = {};
-    for (const card of this.state.series) {
-      enabledQuestions[card.id] = true;
-    }
+    const enabledCards = new Set(this.state.series.map(card => card.id));
 
     const series = [dashcard.card]
       .concat(this.state.series)
@@ -221,16 +162,9 @@ class AddSeriesModal extends Component {
           }}
         >
           <QuestionList
-            questions={questions}
-            badQuestions={badQuestions}
-            enabledQuestions={enabledQuestions}
-            error={error}
+            enabledCards={enabledCards}
             onSelect={this.handleQuestionSelectedChange}
             dashcard={this.props.dashcard}
-            dashcardData={this.props.dashcardData}
-            loadMetadataForQueries={this.handleLoadMetadata}
-            visualization={visualization}
-            isLoadingMetadata={this.state.isLoadingMetadata}
           />
         </div>
       </div>
@@ -238,12 +172,6 @@ class AddSeriesModal extends Component {
   }
 }
 
-export default _.compose(
-  Questions.loadList({ query: { f: "all" } }),
-  connect(
-    (state, ownProps) => ({
-      questions: getQuestions(state, ownProps),
-    }),
-    { loadMetadataForQueries },
-  ),
-)(AddSeriesModal);
+export default _.compose(connect(null, { loadMetadataForQueries }))(
+  AddSeriesModal,
+);
